requirements.txt
aniso8601==2.0.0
appdirs==1.4.3
astroid==1.6.0
backports.functools-lru-cache==1.4
certifi==2017.11.5
chardet==3.0.4
click==6.7
configparser==3.5.0
DateTime==4.2
eikon==0.1.10
enum34==1.1.6
Flask==0.12.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.6
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.14.0
pandas==0.22.0
pylint==1.8.1
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
singledispatch==3.4.0.3
six==1.11.0
urllib3==1.22
websocket-client==0.46.0
Werkzeug==0.14.1
wrapt==1.10.11
zope.interface==4.4.3
